requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.29.4
azure-identity==1.14.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
docopt==0.6.2
fonttools==4.42.1
frozenlist==1.4.0
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.0
msal==1.24.0
msal-extensions==1.0.0
multidict==6.0.4
num2words==0.5.12
numpy==1.26.0
openai==0.28.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
yarl==1.9.2